requirements.txt
dnspython==1.15.0
Flask==0.12.2
zappa==0.42.2
